+++ postgres/initdb/02-init.sql
@@ -26,7 +26,7 @@ create table versions (
   id uuid primary key,
   file_id uuid not null,
   contents_sha char(56) not null,
-  created_at timestamp with time zone not null default now(),
+  created_at timestamp not null default now(),
   unique (file_id, created_at),
   foreign key (file_id) references files (id) on delete cascade,
   -- constraint name is used in textrepo-app:
@@ -39,7 +39,7 @@ create index version_by_file_id on versions (file_id);
 create table documents(
   id uuid primary key,
   external_id varchar not null unique,
-  created_at timestamp with time zone not null default now()
+  created_at timestamp not null default now()
 );
 
 -- assist FK lookups: add compound index for PK fields in reverse order