+++ scripts/migration/compress-contents/04-rename-columns.sql
@@ -1,0 +1,6 @@
+-- Activate the compressed version by first renaming the
+-- old contents to 'uncompressed' and then renaming the new
+-- contents in 'compressed' column to 'contents'
+
+ALTER TABLE contents RENAME contents TO uncompressed;
+ALTER TABLE contents RENAME compressed TO contents;