+++ scripts/migration/compress-contents/05-destroy-uncompressed.sql
@@ -1,0 +1,1 @@
+ALTER TABLE contents DROP COLUMN uncompressed;