+++ textrepo-app/src/main/java/nl/knaw/huc/resources/rest/VersionContentsResource.java
@@ -60,8 +60,10 @@ public class VersionContentsResource {
     log.debug("Get version contents: versionId={}", versionId);
     var contents = contentsService.getByVersionId(versionId);
     log.debug("Got version contents: {}", contents);
+
+    final byte[] payload = contents.decompressIfCompressedSizeLessThan(2 * 1024 * 1024);
     return Response
-        .ok(contents.getContents(), APPLICATION_OCTET_STREAM)
+        .ok(payload, APPLICATION_OCTET_STREAM)
         .header("Content-Disposition", "attachment;")
         .build();
   }