+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Use timestamp without time zone in versions and documents tables
+- Use (gzip) compressed contents in contents table (migration code available)
 
 ### Changed
 - Use content-type text/plain instead of application/json when putting metadata